requirements.txt
flax
chex
rlax
optax
flashbax
matplotlib
numpy
distrax
pillow
wandb
hydra-core
imageio
moviepy
jax-tqdm
opencv-python
# jaxmarl
supersuit
pettingzoo
numpy==1.26.4
ml-dtypes==0.2.0